authentication deleting users task 60

+++ src/tasks/tasks.service.ts
@@ -38,8 +38,8 @@ export class TasksService {
     //     return task
     // }
 
-  async  deleteTaskById(id:number):Promise<void>{
-        const result=await this.TaskRepository.delete(id)
+  async  deleteTaskById(id:number, user:User):Promise<void>{
+        const result=await this.TaskRepository.delete({id, userId:user.id})
         if(result.affected===0)
         {
             throw new NotFoundException("nahi mila")
